+++ lib/service_layer.rb
@@ -5,4 +5,6 @@ require 'service_layer/version'
 # Provides an easy way to write objects responsible for encapsulating domain
 # layer.
 module ServiceLayer
+  autoload :Base, 'service_layer/base'
+  autoload :Response, 'service_layer/response'
 end