+++ spec/service_layer/response/base_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+RSpec.describe ServiceLayer::Response::Base do
+  subject(:response) { described_class.new }
+
+  describe '#success?' do
+    it { expect { response.success? }.to raise_error NotImplementedError }
+  end
+end