+++ lib/service_layer.rb
@@ -6,5 +6,6 @@ require 'service_layer/version'
 # layer.
 module ServiceLayer
   autoload :Base, 'service_layer/base'
+  autoload :Command, 'service_layer/command'
   autoload :Response, 'service_layer/response'
 end