+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] (2017-11-25)
+
 ### Added
 
 * Add explicit `#failure?` on `Response`. Return the opposite of `#success?`.
@@ -23,4 +25,5 @@ method. This makes it possible to know the state of success of an execution.
 `#perform` method and provides a delegation of the execution logic of the
 command.
 
-[Unreleased]: https://github.com/gemologist/service_layer/compare/v0.1.0...master
+[Unreleased]: https://github.com/gemologist/service_layer/compare/v0.2.0...master
+[0.2.0]: https://github.com/gemologist/service_layer/compare/v0.1.0...v0.2.0